requirements.txt
# Django and core dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7

# Database
mysqlclient==2.2.0
django-mysql==4.12.0

# Redis and channels
redis==5.0.1
django-redis==5.4.0
channels==4.0.0
channels-redis==4.1.0

# Background tasks
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Utilities
Pillow==10.1.0
python-dotenv==1.0.0
gunicorn==21.2.0

